[C8] Split sketchconstraint.go by constraint family (#2230)
model/exchange/translators/inventor/ipt/sketchconstraint.go was 921 LOC,
1.8x the 500-line ceiling. Split the .ipt sketch-constraint reader by
constraint family over the shared record decoding; each file is now under
the ceiling:

- sketchconstraint.go (233): shared record decoding + geometry predicates —
  collectRawCons, vertexCoords/lineEndpoints/circleByEntityID (the entity
  maps both families resolve through) and the line predicates (parallel/
  perpendicular/collinear/lengthsEqual/isHorizontal/…).
- sketchconstraint_geometric.go (360): the value-free relations — geometric
  (H/V/parallel/perp/collinear/equal-length/midpoint), circle concentric/
  equal-radius, ground, symmetry, line-circle tangent, point-on-line.
- sketchconstraint_dimensional.go (360): the valued dimensions — offset,
  angle, radius/diameter, distance, revolve radius, axial length.

Behaviour-preserving: extracted by exact line ranges (verified every dropped
line is a blank separator and no function is duplicated), so each moved const/
type/function body is byte-identical; imports narrowed per file. The .ipt file
has no "pattern"-family constraint code, so the split is geometric +
dimensional over shared. The ipt corpus passes before and after (GOWORK=off,
separate module); module lint issue count unchanged (32, the pre-existing
funlen/gocyclo/gocritic on the Decode* functions travel with them). Net delta
of recognizers / tolerance constants / fallback sites / type assertions: 0.

Part of #2183 (C8: structure/size limits). Closes #2230.

(cherry picked from commit 5c607c9c3d791d1880ce1a021032355ef0a3d3af)
