+++ model/exchange/translators/inventor/ipt/sketchconstraint.go
@@ -39,31 +39,6 @@ const (
 	emptyListMark   = 0x00000010 // the +32 empty-extra-list marker of a plain two-ref constraint node
 )
 
-// GeoKind is a decoded geometric-constraint type.
-type GeoKind int
-
-const (
-	GeoHorizontal GeoKind = iota
-	GeoVertical
-	GeoParallel
-	GeoPerpendicular
-	GeoCollinear
-	GeoEqualLength
-	GeoMidpoint
-	GeoPointOnLine
-	GeoConcentric
-	GeoEqualRadius
-)
-
-// GeoConstraint is a decoded geometric constraint with its lines resolved to endpoint coordinates,
-// so the translator can bind it to the emitted sketch by coordinate. L2 is unused for H/V; Pt is
-// set only for point-on-geometry constraints (midpoint), where it is the pinned point's coordinate.
-type GeoConstraint struct {
-	Kind   GeoKind
-	L1, L2 [2]Point2D
-	Pt     Point2D
-}
-
 // rawCon is a decoded constraint/dimension node before its references are resolved.
 type rawCon struct {
 	off    int
@@ -91,451 +66,6 @@ func collectRawCons(seg []byte) []rawCon {
 	return out
 }
 
-// DecodeGeometricConstraints returns the sketch's value-free geometric constraints (horizontal,
-// vertical, parallel, perpendicular), each resolved to the coordinates of the line(s) it relates.
-// Coincidences are omitted (reproduced by shared points), as are dimensions (their value lives in a
-// parameter node — decoded separately). A constraint whose lines don't resolve, or whose geometry
-// matches neither expected relation, is dropped rather than guessed.
-func DecodeGeometricConstraints(seg []byte) []GeoConstraint {
-	vc := vertexCoords(seg)
-	le := lineEndpoints(seg, vc)
-	var out []GeoConstraint
-	for _, c := range collectRawCons(seg) {
-		switch c.disc {
-		case axisAlignDisc:
-			l, ok := le[c.r1&^refBit]
-			if !ok || c.r1&refBit == 0 {
-				continue
-			}
-			if isHorizontal(l) {
-				out = append(out, GeoConstraint{Kind: GeoHorizontal, L1: l})
-			} else if isVertical(l) {
-				out = append(out, GeoConstraint{Kind: GeoVertical, L1: l})
-			}
-		case lineRelateDisc:
-			l1, ok1 := le[c.r1&^refBit]
-			l2, ok2 := le[c.r2&^refBit]
-			if !ok1 || !ok2 || c.r1&refBit == 0 || c.r2&refBit == 0 {
-				continue
-			}
-			// Collinear shares this discriminator with parallel/perpendicular (all "line relate"
-			// two lines) and IS parallel, so it must be tested first: two collinear lines lie on one
-			// infinite line, which parallel-but-offset lines do not.
-			if linesCollinear(l1, l2) {
-				out = append(out, GeoConstraint{Kind: GeoCollinear, L1: l1, L2: l2})
-			} else if linesParallel(l1, l2) {
-				out = append(out, GeoConstraint{Kind: GeoParallel, L1: l1, L2: l2})
-			} else if linesPerpendicular(l1, l2) {
-				out = append(out, GeoConstraint{Kind: GeoPerpendicular, L1: l1, L2: l2})
-			}
-		case coincidenceDisc:
-			// A coincidence node whose BOTH references are lines (not the usual line↔endpoint) is an
-			// equal-length constraint. lineEndpoints already ignores it (it needs one ref to be a
-			// point), so this read doesn't disturb endpoint resolution. Emitted only when the lengths
-			// actually match, so it never implies geometry the sketch doesn't already have.
-			l1, ok1 := le[c.r1&^refBit]
-			l2, ok2 := le[c.r2&^refBit]
-			if ok1 && ok2 && c.r1&refBit != 0 && c.r2&refBit != 0 && lengthsEqual(l1, l2) {
-				out = append(out, GeoConstraint{Kind: GeoEqualLength, L1: l1, L2: l2})
-			}
-		case midpointDisc:
-			// A disc-0 node whose first ref resolves to a line is a MIDPOINT constraint: a sketch
-			// point pinned to the line's midpoint. Radius/diameter dimensions share disc 0 but their
-			// first ref is the 0x10 sentinel (not a ref), so lineEndpoints won't resolve it — the
-			// line gate cleanly excludes them. The pinned point's coordinate is the line's geometric
-			// midpoint, COMPUTED from the resolved line rather than read from the (offset, hard to
-			// resolve) point reference; the apply step binds it only when a sketch point actually sits
-			// there, so a stray disc-0 node can't invent a constraint.
-			if c.r1&refBit == 0 {
-				continue
-			}
-			if l, ok := le[c.r1&^refBit]; ok {
-				out = append(out, GeoConstraint{Kind: GeoMidpoint, L1: l, Pt: midpointOf(l)})
-			}
-		}
-	}
-	out = append(out, decodePointOnLine(seg, vc)...)
-	return out
-}
-
-// CircleRelation is a decoded circle↔circle constraint (concentric or equal-radius), resolved to
-// both circles' centres and radii so the translator can bind them by coordinate.
-type CircleRelation struct {
-	Kind   GeoKind // GeoConcentric or GeoEqualRadius
-	C1, C2 Point2D // centres
-	R1, R2 float64 // radii
-}
-
-// DecodeCircleRelations returns the sketch's concentric and equal-radius constraints. Both are
-// stored as a coincidence node (t44 0x3e) whose two references resolve to CIRCLES (each named by
-// its entity id = centre ref + 1, via circleByEntityID) — the same shape equal-length uses for two
-// lines. Concentric and equal-radius share the discriminator and are told apart by geometry: a
-// concentric pair shares a centre (different radii); an equal-radius pair shares a radius
-// (different centres). Emitted only when the geometry actually exhibits the relation, so a stray
-// node never yields a spurious constraint and nothing moves.
-func DecodeCircleRelations(seg []byte) []CircleRelation {
-	circ := circleByEntityID(seg, vertexCoords(seg))
-	var out []CircleRelation
-	for _, c := range collectRawCons(seg) {
-		if c.disc != coincidenceDisc || c.r1&refBit == 0 || c.r2&refBit == 0 {
-			continue
-		}
-		a, ok1 := circ[c.r1&^refBit]
-		b, ok2 := circ[c.r2&^refBit]
-		if !ok1 || !ok2 {
-			continue
-		}
-		sameCentre := samePoint2D(a.inline, b.inline)
-		sameRadius := math.Abs(a.radius-b.radius) < 1e-4
-		switch {
-		case sameCentre && !sameRadius:
-			out = append(out, CircleRelation{Kind: GeoConcentric, C1: a.inline, C2: b.inline, R1: a.radius, R2: b.radius})
-		case sameRadius && !sameCentre:
-			out = append(out, CircleRelation{Kind: GeoEqualRadius, C1: a.inline, C2: b.inline, R1: a.radius, R2: b.radius})
-		}
-	}
-	return out
-}
-
-// OffsetDim is a decoded offset (distance-from-line) dimension: the perpendicular distance from a
-// point to a reference line. Two Inventor forms decode to this: a point-to-line offset, and a
-// line-to-line offset between two PARALLEL lines (reduced here to the reference line plus one
-// endpoint of the other line, whose perpendicular distance is the line-to-line separation). The
-// value is the current perpendicular distance, so applying it never moves geometry.
-type OffsetDim struct {
-	Line  [2]Point2D // the reference line (r2)
-	Pt    Point2D    // the dimensioned point, or an endpoint of the parallel line
-	Value float64    // perpendicular distance (cm)
-}
-
-// DecodeOffsetDimensions returns the sketch's offset (distance-from-line) dimensions. The node has
-// the same head as an angle dimension — first ref = the 0x10 list sentinel, second ref (+40) = a
-// line — but its t44 (+44) reference is either a POINT (point-to-line offset) or a LINE that is
-// PARALLEL to the second (line-to-line offset). An angle dimension's t44 line is NON-parallel, so
-// the angle decoder (which drops near-parallel pairs) and this one are disjoint. The value is the
-// current perpendicular distance, self-validating.
-func DecodeOffsetDimensions(seg []byte) []OffsetDim {
-	vc := vertexCoords(seg)
-	le := lineEndpoints(seg, vc)
-	var out []OffsetDim
-	for _, c := range collectRawCons(seg) {
-		if c.r1 != emptyListMark || c.r2&refBit == 0 || c.disc&refBit == 0 {
-			continue
-		}
-		line, isLine := le[c.r2&^refBit]
-		if !isLine {
-			continue
-		}
-		if pt, ok := vc[c.disc&^refBit]; ok {
-			if d := pointLineDistance(pt, line); d > 1e-4 {
-				out = append(out, OffsetDim{Line: line, Pt: pt, Value: d})
-			}
-		} else if other, ok := le[c.disc&^refBit]; ok && linesParallel(line, other) {
-			// line-to-line offset: an endpoint of the other (parallel) line to the reference line.
-			if d := pointLineDistance(other[0], line); d > 1e-4 {
-				out = append(out, OffsetDim{Line: line, Pt: other[0], Value: d})
-			}
-		}
-	}
-	return out
-}
-
-// pointLineDistance returns the perpendicular distance from a point to a segment's infinite line
-// (falling back to point-to-endpoint distance for a degenerate segment).
-func pointLineDistance(p Point2D, l [2]Point2D) float64 {
-	dx, dy := l[1].X-l[0].X, l[1].Y-l[0].Y
-	length := math.Hypot(dx, dy)
-	if length < 1e-9 {
-		return math.Hypot(p.X-l[0].X, p.Y-l[0].Y)
-	}
-	return math.Abs((p.X-l[0].X)*dy-(p.Y-l[0].Y)*dx) / length
-}
-
-// AngleDim is a decoded angle dimension between two lines, resolved to both lines' endpoints and
-// the unsigned angle between them in degrees ([0,180]). The value is the current geometric angle,
-// so applying AddAngle(l1, l2, "<deg> deg") pins the angle without moving geometry.
-type AngleDim struct {
-	L1, L2  [2]Point2D
-	Degrees float64
-}
-
-// DecodeAngleDimensions returns the sketch's two-line angle dimensions. An angle-dimension node has
-// its first reference = the 0x10 list sentinel (its label sits at +32, so the first entity ref is
-// the empty-list marker, like a radius dim) and its second reference (+40) AND its t44 word (+44)
-// BOTH resolve to lines — the two lines it dimensions. A distance dimension's t44 references a text
-// point; a symmetry's first two refs are points; a radius/tangent's line-bearing ref is a circle —
-// so none of those pass the two-lines gate. The value is the current unsigned angle between the
-// lines, so the dimension reproduces without moving geometry. A near-parallel pair (angle ≈ 0/180)
-// is dropped: you don't dimension the angle between parallel lines, and it is ill-posed to solve.
-func DecodeAngleDimensions(seg []byte) []AngleDim {
-	vc := vertexCoords(seg)
-	le := lineEndpoints(seg, vc)
-	var out []AngleDim
-	for _, c := range collectRawCons(seg) {
-		if c.r1 != emptyListMark || c.r2&refBit == 0 || c.disc&refBit == 0 {
-			continue
-		}
-		l1, ok1 := le[c.r2&^refBit]
-		l2, ok2 := le[c.disc&^refBit]
-		if !ok1 || !ok2 {
-			continue
-		}
-		deg := lineAngleDegrees(l1, l2)
-		if deg < 1e-2 || deg > 180-1e-2 {
-			continue // parallel/anti-parallel — not an angle dimension
-		}
-		out = append(out, AngleDim{L1: l1, L2: l2, Degrees: deg})
-	}
-	return out
-}
-
-// lineAngleDegrees returns the unsigned angle in degrees ([0,180]) between two segments' directions,
-// using the same measure AddAngle solves against: atan2(|cross|, dot).
-func lineAngleDegrees(a, b [2]Point2D) float64 {
-	ax, ay := a[1].X-a[0].X, a[1].Y-a[0].Y
-	bx, by := b[1].X-b[0].X, b[1].Y-b[0].Y
-	return math.Atan2(math.Abs(ax*by-ay*bx), ax*bx+ay*by) * 180 / math.Pi
-}
-
-// RadiusDim is a decoded radius or diameter dimension of a circle or an arc, resolved to the
-// curve's geometry so the translator can bind it by coordinate. Radius and diameter are
-// byte-identical in the node (both pin the same degree of freedom — the curve's radius), so both
-// decode to this and apply as a radius dimension of the curve's OWN radius; the value equals the
-// current radius, so applying it removes a DOF without moving geometry. The label form (r vs ⌀) is
-// not recoverable and not needed for DOF parity.
-type RadiusDim struct {
-	Center Point2D
-	Radius float64
-	Arc    bool    // true → bind an arc (matched by centre + radius), false → a circle
-	Start  Point2D // arc endpoints (zero for a circle)
-	End    Point2D
-}
-
-// DecodeRadiusDimensions returns the sketch's radius/diameter dimensions. A radius-dimension node
-// has t44 == 0 (shared with midpoint) but its first reference is the 0x10 list sentinel — a
-// midpoint's first reference is a LINE — and its SECOND reference resolves to a circle or arc
-// entity id (circle = centre ref + 1; arc = its highest point ref + 1). The value is the curve's
-// own decoded radius, so the dimension reproduces exactly. A node whose second reference resolves
-// to neither a circle nor an arc is dropped rather than guessed.
-//
-// Radius, diameter, AND arc-length dimensions all share this exact node shape at the constraint
-// level (t44 == 0, 0x10 sentinel, curve ref) — they are byte-identical apart from positional ids;
-// their type lives only in a deep dimension sub-structure this decoder does not read. All three are
-// therefore decoded here as a radius of the curve's own radius. This is faithful for DOF parity:
-// each removes exactly one degree of freedom and, being self-validated against the current radius,
-// moves no geometry — so an arc-length dimension reproduced as an arc-radius dimension yields the
-// same DOF and the same solved geometry. The only unrecoverable detail is the label (r / ⌀ / arc
-// length), which does not affect parity. (Same reasoning as radius-vs-diameter, which are likewise
-// indistinguishable.)
-func DecodeRadiusDimensions(seg []byte) []RadiusDim {
-	vc := vertexCoords(seg)
-	circ := circleByEntityID(seg, vc)
-	arcs := arcByEntityID(seg, vc)
-	var out []RadiusDim
-	for _, c := range collectRawCons(seg) {
-		if c.disc != 0 || c.r1 != emptyListMark || c.r2&refBit == 0 {
-			continue
-		}
-		id := c.r2 &^ refBit
-		if ce, ok := circ[id]; ok {
-			out = append(out, RadiusDim{Center: ce.inline, Radius: ce.radius})
-		} else if a, ok := arcs[id]; ok {
-			out = append(out, RadiusDim{Center: a.Center, Radius: a.Radius, Arc: true, Start: a.Start, End: a.End})
-		}
-	}
-	return out
-}
-
-// arcByEntityID maps each arc's ENTITY id to the arc with its centre and endpoints resolved and its
-// inline radius. Inventor allocates the arc entity right after its centre/start/end point nodes, so
-// the entity id is the highest of those three point references + 1 — the same "+1 after the
-// constituent point(s)" rule circleByEntityID uses for a circle's single centre. This is the id a
-// radius dimension uses to name the arc.
-func arcByEntityID(seg []byte, vc map[uint32]Point2D) map[uint32]Arc {
-	m := map[uint32]Arc{}
-	for _, it := range collectItems(seg) {
-		if it.arc == nil {
-			continue
-		}
-		a := it.arc
-		arc := Arc{Radius: a.radius}
-		if p, ok := vc[a.center]; ok {
-			arc.Center = p
-		}
-		if p, ok := vc[a.start]; ok {
-			arc.Start = p
-		}
-		if p, ok := vc[a.end]; ok {
-			arc.End = p
-		}
-		m[maxRef(a.center, a.start, a.end)+1] = arc
-	}
-	return m
-}
-
-func maxRef(a, b, c uint32) uint32 {
-	m := a
-	if b > m {
-		m = b
-	}
-	if c > m {
-		m = c
-	}
-	return m
-}
-
-// GroundKind selects which entity a decoded ground constraint fixes.
-type GroundKind int
-
-const (
-	GroundLine GroundKind = iota
-	GroundCircle
-	GroundPoint
-)
-
-// GroundConstraint is a decoded "fully fix this geometry" (Inventor Ground): one entity frozen at
-// its current position. It carries the resolved entity by coordinate; only the field selected by
-// Kind is meaningful. Grounding pins the entity where it already sits, so applying it removes
-// degrees of freedom WITHOUT moving anything.
-type GroundConstraint struct {
-	Kind   GroundKind
-	Line   [2]Point2D // GroundLine
-	Center Point2D    // GroundCircle
-	Radius float64    // GroundCircle
-	Pt     Point2D    // GroundPoint
-}
-
-// DecodeGroundConstraints returns the sketch's ground constraints. A ground node's t44 is
-// groundDisc and its FIRST reference names the grounded entity (its second word is a non-ref
-// count, not an entity). The entity is resolved as a line, a circle (by entity id = centre + 1),
-// or a bare point — whichever the reference matches — so the translator can bind it by coordinate.
-// A node whose reference resolves to none of these is dropped rather than guessed.
-func DecodeGroundConstraints(seg []byte) []GroundConstraint {
-	vc := vertexCoords(seg)
-	le := lineEndpoints(seg, vc)
-	circ := circleByEntityID(seg, vc)
-	var out []GroundConstraint
-	for _, c := range collectRawCons(seg) {
-		if c.disc != groundDisc || c.r1&refBit == 0 {
-			continue
-		}
-		id := c.r1 &^ refBit
-		switch {
-		case hasLine(le, id):
-			out = append(out, GroundConstraint{Kind: GroundLine, Line: le[id]})
-		case hasCircle(circ, id):
-			ce := circ[id]
-			out = append(out, GroundConstraint{Kind: GroundCircle, Center: ce.inline, Radius: ce.radius})
-		case hasPoint(vc, id):
-			out = append(out, GroundConstraint{Kind: GroundPoint, Pt: vc[id]})
-		}
-	}
-	return out
-}
-
-func hasLine(m map[uint32][2]Point2D, id uint32) bool  { _, ok := m[id]; return ok }
-func hasCircle(m map[uint32]circleEnt, id uint32) bool { _, ok := m[id]; return ok }
-func hasPoint(m map[uint32]Point2D, id uint32) bool    { _, ok := m[id]; return ok }
-
-// SymmetryConstraint is a decoded symmetry: two points mirror-symmetric about an axis line. The
-// two symmetric points sit at the usual +36/+40 references; the axis line is a THIRD reference at
-// +44 (where a plain two-ref constraint carries a small discriminator — here its high bit is set,
-// naming the axis). Resolved to coordinates and self-validated (each point reflects onto the other
-// across the axis) so a coincidentally-shaped node never yields a spurious constraint.
-type SymmetryConstraint struct {
-	P1, P2 Point2D
-	Axis   [2]Point2D
-}
-
-// DecodeSymmetryConstraints returns the sketch's symmetry constraints. Gate: the +32 empty-list
-// sentinel (a plain two-ref layout — a tangent puts a line ref there instead), both entity refs
-// resolve to POINTS, and the t44 word is itself a reference resolving to the axis LINE. The last
-// point distinguishes symmetry from a point-to-point distance dimension, whose t44 also has its
-// high bit set but references a text/label point (not a line). Finally the geometry must actually
-// be symmetric (pointsSymmetric), so a stray node can't invent a constraint and nothing moves.
-func DecodeSymmetryConstraints(seg []byte) []SymmetryConstraint {
-	vc := vertexCoords(seg)
-	le := lineEndpoints(seg, vc)
-	var out []SymmetryConstraint
-	for _, c := range collectRawCons(seg) {
-		if c.r1&refBit == 0 || c.r2&refBit == 0 || c.disc&refBit == 0 ||
-			binary.LittleEndian.Uint32(seg[c.off+32:]) != emptyListMark {
-			continue
-		}
-		axis, isLine := le[c.disc&^refBit]
-		if !isLine {
-			continue
-		}
-		p1, ok1 := vc[c.r1&^refBit]
-		p2, ok2 := vc[c.r2&^refBit]
-		if ok1 && ok2 && pointsSymmetric(p1, p2, axis) {
-			out = append(out, SymmetryConstraint{P1: p1, P2: p2, Axis: axis})
-		}
-	}
-	return out
-}
-
-// pointsSymmetric reports whether p1 and p2 are mirror images across the infinite axis line
-// (reflecting p1 through its perpendicular foot on the axis lands on p2).
-func pointsSymmetric(p1, p2 Point2D, axis [2]Point2D) bool {
-	ax, ay := axis[1].X-axis[0].X, axis[1].Y-axis[0].Y
-	l2 := ax*ax + ay*ay
-	if l2 < 1e-12 {
-		return false
-	}
-	tt := ((p1.X-axis[0].X)*ax + (p1.Y-axis[0].Y)*ay) / l2
-	fx, fy := axis[0].X+tt*ax, axis[0].Y+tt*ay
-	rx, ry := 2*fx-p1.X, 2*fy-p1.Y
-	return math.Abs(rx-p2.X) < 1e-4 && math.Abs(ry-p2.Y) < 1e-4
-}
-
-// TangentConstraint is a decoded line↔circle tangent constraint, resolved to the line's endpoints
-// and the circle's centre and radius so the translator can bind it by coordinate.
-type TangentConstraint struct {
-	Line   [2]Point2D
-	Center Point2D
-	Radius float64
-}
-
-// DecodeTangentConstraints returns the sketch's line↔circle tangents. A tangent node has a
-// different layout from the two-ref geometric constraints: its List6 map is non-empty, so the
-// LINE reference sits at +32 (a plain two-ref constraint holds the 0x10 sentinel there) and the
-// CIRCLE reference is the +44 word (which doubles as the discriminator, high bit set). The circle
-// is named by its ENTITY id = centre-point ref + 1 (circleByEntityID). Emitted only when the line
-// is actually tangent to the circle (perpendicular distance from centre == radius), so a
-// coincidentally-shaped node never yields a spurious constraint and no geometry moves.
-func DecodeTangentConstraints(seg []byte) []TangentConstraint {
-	vc := vertexCoords(seg)
-	le := lineEndpoints(seg, vc)
-	circ := circleByEntityID(seg, vc)
-	var out []TangentConstraint
-	seen := map[[6]int64]bool{}
-	for i := 0; i+48 <= len(seg); i++ {
-		if binary.LittleEndian.Uint32(seg[i:]) != dcNodeTag ||
-			binary.LittleEndian.Uint32(seg[i+8:]) != nullRef ||
-			binary.LittleEndian.Uint32(seg[i+16:]) != constraintHdr {
-			continue
-		}
-		lref := binary.LittleEndian.Uint32(seg[i+32:])
-		cref := binary.LittleEndian.Uint32(seg[i+conDiscOff:])
-		if lref&refBit == 0 || cref&refBit == 0 {
-			continue
-		}
-		l, ok1 := le[lref&^refBit]
-		c, ok2 := circ[cref&^refBit]
-		if !ok1 || !ok2 || !lineTangentToCircle(l, c.inline, c.radius) {
-			continue
-		}
-		key := [6]int64{r4(l[0].X), r4(l[0].Y), r4(l[1].X), r4(l[1].Y), r4(c.inline.X), r4(c.inline.Y)}
-		if seen[key] {
-			continue
-		}
-		seen[key] = true
-		out = append(out, TangentConstraint{Line: l, Center: c.inline, Radius: c.radius})
-	}
-	return out
-}
-
 // circleByEntityID maps each circle's ENTITY id (its centre-point reference + 1, the id a
 // constraint uses to name the circle) to the circle with its centre resolved.
 func circleByEntityID(seg []byte, vc map[uint32]Point2D) map[uint32]circleEnt {
@@ -551,175 +81,6 @@ func circleByEntityID(seg []byte, vc map[uint32]Point2D) map[uint32]circleEnt {
 		m[ce.ref+1] = ce
 	}
 	return m
-}
-
-// lineTangentToCircle reports whether the segment's infinite line is tangent to the circle: the
-// perpendicular distance from the centre to the line equals the radius.
-func lineTangentToCircle(l [2]Point2D, center Point2D, radius float64) bool {
-	dx, dy := l[1].X-l[0].X, l[1].Y-l[0].Y
-	length := math.Hypot(dx, dy)
-	if length < 1e-9 {
-		return false
-	}
-	dist := math.Abs((center.X-l[0].X)*dy-(center.Y-l[0].Y)*dx) / length
-	return math.Abs(dist-radius) < 1e-4
-}
-
-// DistanceDim is a decoded point-to-point distance dimension: the two sketch points it dimensions
-// and their separation (cm). The value is taken from the geometry (the points' distance), which
-// equals the dimension's stored value for a driving dimension at the solved geometry — so applying
-// AddDistance(A,B,value) reproduces the dimension without moving a point.
-type DistanceDim struct {
-	A, B  Point2D
-	Value float64
-}
-
-// DecodeDistanceDimensions returns the sketch's point-to-point distance dimensions. A distance
-// dimension is the one constraint node whose BOTH references resolve to distinct sketch points: a
-// coincidence's first reference is a line, parallel/perpendicular reference two lines, an
-// axis-align references a line, and a radius uses the 0x10 sentinel — so none of those pass the
-// two-points gate, leaving only the distance dimensions. (Aligned dimensions whose second endpoint
-// isn't captured, and distance-from-line dimensions, are not covered here.)
-func DecodeDistanceDimensions(seg []byte) []DistanceDim {
-	vc := vertexCoords(seg)
-	le := lineEndpoints(seg, vc)
-	var out []DistanceDim
-	seen := map[[4]int64]bool{}
-	for _, c := range collectRawCons(seg) {
-		// A point-to-point distance dimension's t44 is itself a reference (its high bit is set — it
-		// points at the dimension's text/label entity), unlike a geometric constraint whose t44 is a
-		// small enum (coincidence 0x3e, axis-align 0x01030000, line-relate 0x00400000, radius 0).
-		// Requiring t44 to be a reference keeps a two-point *geometric* relation (align/symmetry) from
-		// being mistaken for a distance dimension.
-		if c.disc&refBit == 0 || c.r1&refBit == 0 || c.r2&refBit == 0 {
-			continue
-		}
-		// A SYMMETRY node has the same shape (high-bit t44, two point refs), but its t44 references
-		// the symmetry AXIS line rather than a text point. So skip when t44 resolves to a line —
-		// otherwise the two symmetric points would be misread as a spurious distance dimension.
-		if _, isAxis := le[c.disc&^refBit]; isAxis {
-			continue
-		}
-		p1, ok1 := vc[c.r1&^refBit]
-		p2, ok2 := vc[c.r2&^refBit]
-		if !ok1 || !ok2 {
-			continue
-		}
-		d := math.Hypot(p1.X-p2.X, p1.Y-p2.Y)
-		if d < 1e-3 {
-			continue // coincident points are not a distance dimension
-		}
-		key := distKey(p1, p2)
-		if seen[key] {
-			continue
-		}
-		seen[key] = true
-		out = append(out, DistanceDim{A: p1, B: p2, Value: d})
-	}
-	return out
-}
-
-// DecodeRevolveRadii returns the x-distances at which a revolve's radius dimensions pin the profile
-// edges from the x=0 centreline. A revolve radius node (t44=0x01150000) carries an inline value; we
-// accept it only when it EQUALS the x-position of an actual vertical profile edge (so the value is a
-// genuine radius, not a coordinate leak) and return that x. Applied as a horizontal distance from
-// the centreline, the value is exactly the edge's current x — geometry-safe. Reuniting the
-// centreline into the profile sketch (ReuniteRevolveAxis) is what lets these bind: the x=0 axis and
-// the x=V edge then live in one sketch.
-func DecodeRevolveRadii(seg []byte) []float64 {
-	if !HasRevolve(seg) {
-		return nil // radius-from-centreline dimensions are meaningful only on a revolve profile
-	}
-	vc := vertexCoords(seg)
-	le := lineEndpoints(seg, vc)
-	edgeX := map[int64]bool{}
-	for _, l := range le {
-		if math.Abs(l[0].X-l[1].X) < geoEps && math.Abs(l[0].X) > 1e-3 {
-			edgeX[r4(l[0].X)] = true
-		}
-	}
-	var out []float64
-	seen := map[int64]bool{}
-	for _, c := range collectRawCons(seg) {
-		if c.disc != axisRadiusDisc {
-			continue
-		}
-		for o := c.off + conDiscOff; o+8 <= c.off+140 && o+8 <= len(seg); o += 4 {
-			v := math.Float64frombits(binary.LittleEndian.Uint64(seg[o:]))
-			if edgeX[r4(v)] && !seen[r4(v)] {
-				seen[r4(v)] = true
-				out = append(out, v)
-				break
-			}
-		}
-	}
-	return out
-}
-
-// AxialLength is a decoded axial (step-length) dimension of a revolve profile: the vertical gap
-// between two horizontal profile edges. It is accepted only when the gap equals a model-parameter
-// value — the modeller drives each step length by a parameter, so a gap that matches one is a real
-// dimension, not an incidental alignment. Applied as a vertical distance, the value is the edges'
-// current separation, so geometry never moves.
-type AxialLength struct {
-	Y1, Y2 float64
-	Value  float64
-}
-
-// DecodeAxialLengths returns the revolve profile's axial step-length dimensions: the vertical gaps
-// between ADJACENT horizontal edges whose size matches a model parameter (the driving dimension).
-func DecodeAxialLengths(seg []byte) []AxialLength {
-	if !HasRevolve(seg) {
-		return nil // step-length dimensions are meaningful only on a revolve profile
-	}
-	le := lineEndpoints(seg, vertexCoords(seg))
-	var ys []float64
-	seen := map[int64]bool{}
-	for _, l := range le {
-		if math.Abs(l[0].Y-l[1].Y) < geoEps && !seen[r4(l[0].Y)] {
-			seen[r4(l[0].Y)] = true
-			ys = append(ys, l[0].Y)
-		}
-	}
-	sort.Float64s(ys)
-	params := paramValueSet(seg)
-	var out []AxialLength
-	for i := 1; i < len(ys); i++ {
-		gap := ys[i] - ys[i-1]
-		if gap > 1e-3 && params[r4(gap)] {
-			out = append(out, AxialLength{Y1: ys[i-1], Y2: ys[i], Value: gap})
-		}
-	}
-	return out
-}
-
-// paramValueSet is the set of model-parameter values (rounded), used to certify that a decoded gap
-// is a driven dimension rather than an incidental one.
-func paramValueSet(seg []byte) map[int64]bool {
-	m := map[int64]bool{}
-	for i := 0; i+20 <= len(seg); i++ {
-		if binary.LittleEndian.Uint32(seg[i:]) != dcNodeTag || binary.LittleEndian.Uint32(seg[i+8:]) != nullRef {
-			continue
-		}
-		n := int(binary.LittleEndian.Uint32(seg[i+16:]))
-		if n < 1 || n > 64 || i+20+n > len(seg) {
-			continue
-		}
-		if v, ok := firstNonZeroDouble(seg, i+20+n, 40); ok && v > 1e-3 && v < 100 {
-			m[r4(v)] = true
-		}
-	}
-	return m
-}
-
-// distKey is an order-independent rounded key for a point pair, so a doubly-recorded dimension
-// de-duplicates.
-func distKey(a, b Point2D) [4]int64 {
-	ax, ay, bx, by := r4(a.X), r4(a.Y), r4(b.X), r4(b.Y)
-	if ax > bx || (ax == bx && ay > by) {
-		ax, ay, bx, by = bx, by, ax, ay
-	}
-	return [4]int64{ax, ay, bx, by}
 }
 
 func r4(v float64) int64 { return int64(math.Round(v * 1e4)) }
@@ -869,53 +230,4 @@ func lineEndpoints(seg []byte, vc map[uint32]Point2D) map[uint32][2]Point2D {
 		}
 	}
 	return out
-}
-
-// extremePair returns the two points farthest apart in a collinear set — a line's endpoints.
-func extremePair(ps []Point2D) [2]Point2D {
-	ai, bi, best := 0, 1, -1.0
-	for i := range ps {
-		for j := i + 1; j < len(ps); j++ {
-			if d := math.Hypot(ps[i].X-ps[j].X, ps[i].Y-ps[j].Y); d > best {
-				best, ai, bi = d, i, j
-			}
-		}
-	}
-	return [2]Point2D{ps[ai], ps[bi]}
-}
-
-// decodePointOnLine returns the sketch's point-on-line constraints: a curve vertex pinned onto a
-// line's INTERIOR by a 0x3e coincidence node (a coincidence at an endpoint is a plain corner, not a
-// point-on-line). Each is real — a decoded coincidence node whose point resolves to a vertex lying
-// strictly between the line's endpoints — so applying it moves no geometry. (A point-on-line on a
-// STANDALONE point isn't covered: a standalone point isn't a curve vertex, so its arbitrary
-// position along the line can't be resolved.)
-func decodePointOnLine(seg []byte, vc map[uint32]Point2D) []GeoConstraint {
-	var out []GeoConstraint
-	for _, ps := range coincidenceGroups(seg, vc) {
-		if len(ps) < 3 {
-			continue
-		}
-		e := extremePair(ps)
-		for _, p := range ps {
-			if !samePoint2D(p, e[0]) && !samePoint2D(p, e[1]) && onSegmentInterior(p, e[0], e[1]) {
-				out = append(out, GeoConstraint{Kind: GeoPointOnLine, L1: e, Pt: p})
-			}
-		}
-	}
-	return out
-}
-
-// onSegmentInterior reports whether p lies on segment a–b strictly between the endpoints.
-func onSegmentInterior(p, a, b Point2D) bool {
-	dx, dy := b.X-a.X, b.Y-a.Y
-	length := math.Hypot(dx, dy)
-	if length < 1e-6 {
-		return false
-	}
-	if math.Abs((p.X-a.X)*dy-(p.Y-a.Y)*dx)/length > 1e-4 {
-		return false // not on the line
-	}
-	tt := ((p.X-a.X)*dx + (p.Y-a.Y)*dy) / (length * length)
-	return tt > 1e-3 && tt < 1-1e-3
 }
